requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# LLM API clients
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0
ollama>=0.1.0

# Data validation and parsing
pydantic>=2.0.0
jsonschema>=4.20.0

# Async processing
aiohttp>=3.9.0
asyncio>=3.4.3

# Visualization and dashboard
streamlit>=1.31.0
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
